refactor: move all extensions to separate package
