+++ mobile/lib/extensions/collection_extensions.dart
@@ -2,27 +2,6 @@ import 'dart:typed_data';
 
 import 'package:collection/collection.dart';
 
-extension DurationExtension on String {
-  Duration? toDuration() {
-    try {
-      final parts = split(':')
-          .map((e) => double.parse(e).toInt())
-          .toList(growable: false);
-      return Duration(hours: parts[0], minutes: parts[1], seconds: parts[2]);
-    } catch (e) {
-      return null;
-    }
-  }
-
-  double toDouble() {
-    return double.parse(this);
-  }
-
-  int toInt() {
-    return int.parse(this);
-  }
-}
-
 extension ListExtension<E> on List<E> {
   List<E> uniqueConsecutive({
     int Function(E a, E b)? compare,
