--- mobile/lib/utils/capitalize.dart
@@ -1,9 +1,0 @@
-extension StringExtension on String {
-  String capitalize() {
-    return split(" ")
-        .map(
-          (str) => str.isEmpty ? str : str[0].toUpperCase() + str.substring(1),
-        )
-        .join(" ");
-  }
-}
